docs: add caveat about ContextVar in dependencies
As discovered and discussed in #13382, ContextVars set in a
_synchronous_ dependency is not available to the rest of the request due
to the way FastAPI handles that dependency.

The documentation did not make any mention of this, and it was only
after I asked that I discovered this was intended behaviour. So I am
updating the docs to make this clear to others.

Ref: #13382
Signed-off-by: JP-Ellis <josh@jpellis.me>